+++ src/Actions/Email/Action.php
@@ -4,15 +4,18 @@ declare(strict_types=1);
 
 namespace Zkwbbr\WhoopsHelper\Actions\Email;
 
+use Zkwbbr\WhoopsHelper\Actions\ActionInterface;
+use MetaRush\EmailFallback\Builder;
+
 /**
  * Send error to email
  */
-class Action implements \Zkwbbr\WhoopsHelper\Actions\ActionInterface
+class Action implements ActionInterface
 {
     private $siteUrl;
     private $mailBuilder;
 
-    public function __construct(string $siteUrl, \MetaRush\EmailFallback\Builder $mailBuilder)
+    public function __construct(string $siteUrl, Builder $mailBuilder)
     {
         $this->siteUrl = $siteUrl;
         $this->mailBuilder = $mailBuilder;
@@ -24,7 +27,7 @@ class Action implements \Zkwbbr\WhoopsHelper\Actions\ActionInterface
         $this->mailBuilder->setBody('[' . $date . '] ' . $ex->getMessage() . ' on ' . $ex->getFile() . ' (' . $ex->getLine() . ')');
 
         $mailer = $this->mailBuilder->build();
-
         $mailer->sendEmailFallback();
     }
+
 }